requirements.txt
requests
fade
colorama
python-dotenv
customtkinter
Pillow